Adiciona nível de permissão de páginas

+++ mod_cliente/cliente_controller.py
@@ -1,5 +1,5 @@
 # coding: utf-8
-from flask import Blueprint, render_template, request, url_for, redirect
+from flask import Blueprint, render_template, request, url_for, redirect, session
 from mod_login.login_controller import logado
 from mod_cliente.cliente_model import Cliente
 
@@ -12,49 +12,57 @@ bp_cliente = Blueprint('cliente', __name__, url_prefix='/', template_folder='tem
 @bp_cliente.route('/clientes', methods=['GET'])
 @logado
 def lista():
-    cliente = Cliente()
-    clientes = cliente.all()
-    return render_template('lista_cliente.html', lista=clientes)
+    if session['user']['grupo'] == 'admin':
+        cliente = Cliente()
+        clientes = cliente.all()
+        return render_template('lista_cliente.html', lista=clientes)
+    return render_template('403.html'), 403
 
 
 @bp_cliente.route('/cliente', methods=['GET'])
 @logado
 def cadastro_form():
-    # Paǵina de cadastro
-    cliente = Cliente()
-    return render_template('form_cliente.html', cliente=cliente)
+    if session['user']['grupo'] == 'admin':
+        # Paǵina de cadastro
+        cliente = Cliente()
+        return render_template('form_cliente.html', cliente=cliente)
+    return render_template('403.html'), 403
 
 
 @bp_cliente.route('/cliente/<int:clienteid>', methods=['GET'])
 @logado
 def edicao_form(clienteid: int):
-    # Página de edição
     cliente = Cliente()
     cliente.select(clienteid)
-    if cliente.id_cliente == 0:
-        return redirect(url_for('cliente.lista'))
-    return render_template('form_cliente.html', cliente=cliente)
+    if session['user']['grupo'] == 'admin' or session['user']['id_cliente'] == cliente.id_cliente:
+        # Página de edição
+        if cliente.id_cliente == 0:
+            return redirect(url_for('cliente.lista'))
+        return render_template('form_cliente.html', cliente=cliente)
+    return render_template('403.html'), 403
 
 
 @bp_cliente.route('/cliente', methods=['POST'])
 @logado
 def cadastro():
-    # Cadastro via ajax
-    cliente = Cliente()
-    populate_from_request(cliente)
+    if session['user']['grupo'] == 'admin':
+        # Cadastro via ajax
+        cliente = Cliente()
+        populate_from_request(cliente)
 
-    if not Cliente.valid_pass(request.form['senha']):
-        return json_response(message='A senha deve ter pelo menos 4 dígitos', data=[]), 400
+        if not Cliente.valid_pass(request.form['senha']):
+            return json_response(message='A senha deve ter pelo menos 4 dígitos', data=[]), 400
 
-    if cliente.login_exists(cliente.login, 0):
-        return json_response(message='O login já está em uso, utilize outro', data=[]), 400
+        if cliente.login_exists(cliente.login, 0):
+            return json_response(message='O login já está em uso, utilize outro', data=[]), 400
 
-    cliente.senha = Cliente.hash(request.form['senha'])
-    identifier = cliente.insert()
-    if identifier > 0:
-        return json_response(message='Cliente cadastrado!', data=[cliente], redirect=url_for('cliente.lista')), 201
-    else:
-        return json_response(message='Não foi possível cadastrar o cliente', data=[]), 400
+        cliente.senha = Cliente.hash(request.form['senha'])
+        identifier = cliente.insert()
+        if identifier > 0:
+            return json_response(message='Cliente cadastrado!', data=[cliente], redirect=url_for('cliente.lista')), 201
+        else:
+            return json_response(message='Não foi possível cadastrar o cliente', data=[]), 400
+    return json_response(message='Você não tem permissão para realizar esta ação', data=[]), 403
 
 
 @bp_cliente.route('/cliente/<int:clienteid>', methods=['POST', 'PUT'])
@@ -64,24 +72,26 @@ def edicao(clienteid):
     # Verifica se usuário existe
     cliente = Cliente()
     cliente.select(clienteid)
-    if cliente.id_cliente == 0:
-        return json_response(message='Cliente não encontrado!', data=[], redirect=url_for('cliente.lista')), 404
-    
-    populate_from_request(cliente)
+    if session['user']['grupo'] == 'admin' or session['user']['id_cliente'] == cliente.id_cliente:
+        if cliente.id_cliente == 0:
+            return json_response(message='Cliente não encontrado!', data=[], redirect=url_for('cliente.lista')), 404
 
-    if len(request.form['senha']) > 0:
-        if not Cliente.valid_pass(request.form['senha']):
-            return json_response(message='A senha deve ter pelo menos 4 dígitos', data=[]), 400
-        cliente.senha = Cliente.hash(request.form['senha'])
+        populate_from_request(cliente)
 
-    if cliente.login_exists(cliente.login, cliente.id_cliente):
-        return json_response(message='O login já está em uso, utilize outro', data=[]), 400
+        if len(request.form['senha']) > 0:
+            if not Cliente.valid_pass(request.form['senha']):
+                return json_response(message='A senha deve ter pelo menos 4 dígitos', data=[]), 400
+            cliente.senha = Cliente.hash(request.form['senha'])
 
-    rows = cliente.update()
-    if rows > 0:
-        return json_response(message='Cliente atualizado!', data=[cliente]), 200
-    else:
-        return json_response(message='Não foi possível editar o cliente', data=[]), 400
+        if cliente.login_exists(cliente.login, cliente.id_cliente):
+            return json_response(message='O login já está em uso, utilize outro', data=[]), 400
+
+        rows = cliente.update()
+        if rows > 0:
+            return json_response(message='Cliente atualizado!', data=[cliente]), 200
+        else:
+            return json_response(message='Não foi possível editar o cliente', data=[]), 400
+    return json_response(message='Você não tem permissão para realizar esta ação', data=[]), 403
 
 
 @bp_cliente.route('/cliente/<int:clienteid>', methods=['DELETE'])
@@ -91,15 +101,17 @@ def remocao(clienteid):
     # Verifica se usuário existe
     cliente = Cliente()
     cliente.select(clienteid)
-    if cliente.id_cliente == 0:
-        return json_response(message='Cliente não encontrado!', data=[], redirect=url_for('cliente.lista')), 404
-    if cliente.bought():
-        return json_response(message='Cliente com pedidos cadastrados não pode ser removido!', data=[]), 403
-    rows = cliente.delete()
-    if rows > 0:
-        return json_response(message='Cliente removido!', data=[cliente], redirect=url_for('cliente.lista')), 200
-    else:
-        return json_response(message='Não foi possível remover o cliente', data=[]), 400
+    if session['user']['grupo'] == 'admin':
+        if cliente.id_cliente == 0:
+            return json_response(message='Cliente não encontrado!', data=[], redirect=url_for('cliente.lista')), 404
+        if cliente.bought():
+            return json_response(message='Cliente com pedidos cadastrados não pode ser removido!', data=[]), 403
+        rows = cliente.delete()
+        if rows > 0:
+            return json_response(message='Cliente removido!', data=[cliente], redirect=url_for('cliente.lista')), 200
+        else:
+            return json_response(message='Não foi possível remover o cliente', data=[]), 400
+    return json_response(message='Você não tem permissão para realizar esta ação', data=[]), 403
 
 
 @bp_cliente.route('/cliente/busca/<int:clienteid>', methods=['GET'])
@@ -108,11 +120,13 @@ def busca(clienteid: int):
     # Busca por cliente
     cliente = Cliente()
     cliente.select(clienteid)
-    if cliente.id_cliente == 0:
-        return json_response(message='Cliente não encontrado!', data=[]), 404
-    cliente.telefone = helper.telefone(cliente.telefone)
-    cliente.cep = helper.cep(cliente.cep)
-    return json_response(message='Cliente encontrado!', data=[cliente]), 200
+    if session['user']['grupo'] == 'admin' or session['user']['id_cliente'] == cliente.id_cliente:
+        if cliente.id_cliente == 0:
+            return json_response(message='Cliente não encontrado!', data=[]), 404
+        cliente.telefone = helper.telefone(cliente.telefone)
+        cliente.cep = helper.cep(cliente.cep)
+        return json_response(message='Cliente encontrado!', data=[cliente]), 200
+    return json_response(message='Você não tem permissão para realizar esta ação', data=[]), 403
 
 
 def populate_from_request(cliente: Cliente):
